backend/requirements.txt
fastapi
uvicorn[standard]
numpy
opencv-python-headless
Pillow
python-multipart
